Remaning Spatial Module Prep
