+++ packages/editor/src/components/properties/AmbientLightNodeEditor.tsx
@@ -27,7 +27,7 @@ import React from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { useComponent } from '@etherealengine/ecs/src/ComponentFunctions'
-import { AmbientLightComponent } from '@etherealengine/engine/src/scene/components/AmbientLightComponent'
+import { AmbientLightComponent } from '@etherealengine/engine/src/renderer/components/AmbientLightComponent'
 
 import Brightness7Icon from '@mui/icons-material/Brightness7'
