+++ packages/editor/src/components/properties/PointLightNodeEditor.tsx
@@ -27,7 +27,7 @@ import React from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { useComponent } from '@etherealengine/ecs/src/ComponentFunctions'
-import { PointLightComponent } from '@etherealengine/engine/src/scene/components/PointLightComponent'
+import { PointLightComponent } from '@etherealengine/engine/src/renderer/components/PointLightComponent'
 
 import LightbulbIcon from '@mui/icons-material/Lightbulb'
