+++ packages/editor/src/components/properties/SpotLightNodeEditor.tsx
@@ -27,7 +27,7 @@ import React from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { useComponent } from '@etherealengine/ecs/src/ComponentFunctions'
-import { SpotLightComponent } from '@etherealengine/engine/src/scene/components/SpotLightComponent'
+import { SpotLightComponent } from '@etherealengine/engine/src/renderer/components/SpotLightComponent'
 
 import AdjustIcon from '@mui/icons-material/Adjust'
