+++ packages/engine/src/avatar/systems/AnimationSystem.ts
@@ -35,8 +35,8 @@ import { defineQuery } from '@etherealengine/ecs/src/QueryFunctions'
 import { defineSystem } from '@etherealengine/ecs/src/SystemFunctions'
 import { traverseEntityNode } from '@etherealengine/engine/src/transform/components/EntityTree'
 import { VRM } from '@pixiv/three-vrm'
+import { MeshComponent } from '../../renderer/components/MeshComponent'
 import { VisibleComponent } from '../../renderer/components/VisibleComponent'
-import { MeshComponent } from '../../scene/components/MeshComponent'
 import { ModelComponent } from '../../scene/components/ModelComponent'
 import { TransformSystem } from '../../transform/TransformModule'
 import { TransformComponent } from '../../transform/components/TransformComponent'
