+++ packages/engine/src/physics/classes/Physics.test.ts
@@ -41,7 +41,7 @@ import { getMutableState, getState } from '@etherealengine/hyperflux'
 import { ObjectDirection } from '../../common/constants/Axis3D'
 import { createEngine } from '../../initializeEngine'
 import { addObjectToGroup } from '../../renderer/components/GroupComponent'
-import { MeshComponent } from '../../scene/components/MeshComponent'
+import { MeshComponent } from '../../renderer/components/MeshComponent'
 import { TransformComponent } from '../../transform/components/TransformComponent'
 import { computeTransformMatrix } from '../../transform/systems/TransformSystem'
 import { CollisionComponent } from '../components/CollisionComponent'
