+++ packages/engine/src/renderer/components/HemisphereLightComponent.ts
@@ -24,54 +24,62 @@ Ethereal Engine. All Rights Reserved.
 */
 
 import { useEffect } from 'react'
-import { AmbientLight, Color } from 'three'
+import { Color, HemisphereLight } from 'three'
 
 import { defineComponent, useComponent } from '@etherealengine/ecs/src/ComponentFunctions'
 import { useEntityContext } from '@etherealengine/ecs/src/EntityFunctions'
 import { matches } from '../../common/functions/MatchesUtils'
-import { addObjectToGroup, removeObjectFromGroup } from '../../renderer/components/GroupComponent'
+import { addObjectToGroup, removeObjectFromGroup } from './GroupComponent'
 
-export const AmbientLightComponent = defineComponent({
-  name: 'AmbientLightComponent',
-  jsonID: 'ambient-light',
+export const HemisphereLightComponent = defineComponent({
+  name: 'HemisphereLightComponent',
+  jsonID: 'hemisphere-light',
 
   onInit: (entity) => {
-    const light = new AmbientLight()
+    const light = new HemisphereLight()
     return {
       light,
-      // todo, maybe we want to reference light.color instead of creating a new Color?
-      color: new Color(),
+      skyColor: new Color(),
+      groundColor: new Color(),
       intensity: 1
     }
   },
 
   onSet: (entity, component, json) => {
     if (!json) return
-    if (matches.object.test(json.color) && json.color.isColor) component.color.set(json.color)
-    if (matches.string.test(json.color)) component.color.value.set(json.color)
+    if (matches.object.test(json.skyColor) && json.skyColor.isColor) component.skyColor.set(json.skyColor)
+    if (matches.string.test(json.skyColor) || matches.number.test(json.skyColor))
+      component.skyColor.value.set(json.skyColor)
+    if (matches.object.test(json.groundColor) && json.groundColor.isColor) component.groundColor.set(json.groundColor)
+    if (matches.string.test(json.groundColor) || matches.number.test(json.groundColor))
+      component.groundColor.value.set(json.groundColor)
     if (matches.number.test(json.intensity)) component.intensity.set(json.intensity)
   },
 
   toJSON: (entity, component) => {
     return {
-      color: component.color.value,
+      skyColor: component.skyColor.value,
+      groundColor: component.groundColor.value,
       intensity: component.intensity.value
     }
   },
 
   reactor: function () {
     const entity = useEntityContext()
-    const light = useComponent(entity, AmbientLightComponent)
+    const light = useComponent(entity, HemisphereLightComponent)
     useEffect(() => {
       addObjectToGroup(entity, light.light.value)
       return () => {
         removeObjectFromGroup(entity, light.light.value)
       }
     }, [])
+    useEffect(() => {
+      light.light.value.groundColor.set(light.groundColor.value)
+    }, [light.groundColor])
 
     useEffect(() => {
-      light.light.value.color.set(light.color.value)
-    }, [light.color])
+      light.light.value.color.set(light.skyColor.value)
+    }, [light.skyColor])
 
     useEffect(() => {
       light.light.value.intensity = light.intensity.value
