+++ packages/engine/src/scene/components/EnvmapComponent.tsx
@@ -55,11 +55,11 @@ import { SceneState } from '@etherealengine/engine/src/scene/Scene'
 import { AssetLoader } from '../../assets/classes/AssetLoader'
 import { RendererState } from '../../renderer/RendererState'
 import { GroupComponent } from '../../renderer/components/GroupComponent'
+import { MeshComponent } from '../../renderer/components/MeshComponent'
 import { EnvMapSourceType, EnvMapTextureType } from '../constants/EnvMapEnum'
 import { getRGBArray, loadCubeMapTexture } from '../constants/Util'
 import { addError, removeError } from '../functions/ErrorFunctions'
 import { EnvMapBakeComponent, applyBoxProjection } from './EnvMapBakeComponent'
-import { MeshComponent } from './MeshComponent'
 
 const tempColor = new Color()
