+++ packages/engine/src/scene/components/GroundPlaneComponent.ts
@@ -45,9 +45,9 @@ import { CollisionGroups } from '../../physics/enums/CollisionGroups'
 import { getInteractionGroups } from '../../physics/functions/getInteractionGroups'
 import { PhysicsState } from '../../physics/state/PhysicsState'
 import { addObjectToGroup, removeObjectFromGroup } from '../../renderer/components/GroupComponent'
+import { MeshComponent } from '../../renderer/components/MeshComponent'
 import { enableObjectLayer } from '../../renderer/components/ObjectLayerComponent'
 import { ObjectLayers } from '../../renderer/constants/ObjectLayers'
-import { MeshComponent } from './MeshComponent'
 import { SceneAssetPendingTagComponent } from './SceneAssetPendingTagComponent'
 import { SceneObjectComponent } from './SceneObjectComponent'
