+++ packages/engine/src/scene/components/ModelComponent.tsx
@@ -56,13 +56,13 @@ import { autoconvertMixamoAvatar, isAvaturn } from '../../avatar/functions/avata
 import { CameraComponent } from '../../camera/components/CameraComponent'
 import { EngineRenderer } from '../../renderer/WebGLRendererSystem'
 import { GroupComponent, addObjectToGroup } from '../../renderer/components/GroupComponent'
+import { MeshComponent } from '../../renderer/components/MeshComponent'
 import { SourceType } from '../../scene/materials/components/MaterialSource'
 import { removeMaterialSource } from '../../scene/materials/functions/MaterialLibraryFunctions'
 import { addError, removeError } from '../functions/ErrorFunctions'
 import { parseGLTFModel, proxifyParentChildRelationships } from '../functions/loadGLTFModel'
 import { getModelSceneID } from '../functions/loaders/ModelFunctions'
 import { EnvmapComponent } from './EnvmapComponent'
-import { MeshComponent } from './MeshComponent'
 import { ObjectGridSnapComponent } from './ObjectGridSnapComponent'
 import { SceneAssetPendingTagComponent } from './SceneAssetPendingTagComponent'
 import { SceneObjectComponent } from './SceneObjectComponent'
