+++ packages/engine/src/scene/functions/loadGLTFModel.ts
@@ -46,6 +46,7 @@ import { NameComponent } from '../../common/NameComponent'
 import iterateObject3D from '../../common/functions/iterateObject3D'
 import { EngineRenderer } from '../../renderer/WebGLRendererSystem'
 import { GroupComponent, addObjectToGroup } from '../../renderer/components/GroupComponent'
+import { MeshComponent } from '../../renderer/components/MeshComponent'
 import { enableObjectLayer } from '../../renderer/components/ObjectLayerComponent'
 import { VisibleComponent } from '../../renderer/components/VisibleComponent'
 import { ObjectLayers } from '../../renderer/constants/ObjectLayers'
@@ -55,7 +56,6 @@ import { computeTransformMatrix } from '../../transform/systems/TransformSystem'
 import { GLTFLoadedComponent } from '../components/GLTFLoadedComponent'
 import { InstancingComponent } from '../components/InstancingComponent'
 import { MeshBVHComponent } from '../components/MeshBVHComponent'
-import { MeshComponent } from '../components/MeshComponent'
 import { ModelComponent } from '../components/ModelComponent'
 import { SceneObjectComponent } from '../components/SceneObjectComponent'
