+++ packages/engine/src/scene/functions/loaders/VariantFunctions.ts
@@ -10,10 +10,10 @@ import { pathResolver } from '../../../assets/functions/pathResolver'
 import { addOBCPlugin } from '../../../common/functions/OnBeforeCompilePlugin'
 import { isMobile } from '../../../common/functions/isMobile'
 import { GroupComponent, addObjectToGroup, removeObjectFromGroup } from '../../../renderer/components/GroupComponent'
+import { MeshComponent } from '../../../renderer/components/MeshComponent'
 import { TransformComponent } from '../../../transform/components/TransformComponent'
 import { isMobileXRHeadset } from '../../../xr/XRState'
 import { InstancingComponent } from '../../components/InstancingComponent'
-import { MeshComponent } from '../../components/MeshComponent'
 import { ModelComponent } from '../../components/ModelComponent'
 import { VariantComponent, VariantLevel } from '../../components/VariantComponent'
 import getFirstMesh from '../../util/meshUtils'
