+++ packages/engine/src/scene/systems/VariantSystem.tsx
@@ -31,9 +31,9 @@ import { defineSystem } from '@etherealengine/ecs/src/SystemFunctions'
 import { EngineState } from '@etherealengine/engine/src/EngineState'
 import { SceneState } from '@etherealengine/engine/src/scene/Scene'
 import { Not } from 'bitecs'
+import { MeshComponent } from '../../renderer/components/MeshComponent'
 import { TransformComponent } from '../../transform/components/TransformComponent'
 import { InstancingComponent } from '../components/InstancingComponent'
-import { MeshComponent } from '../components/MeshComponent'
 import { ModelComponent } from '../components/ModelComponent'
 import { VariantComponent } from '../components/VariantComponent'
 import { setInstancedMeshVariant, setMeshVariant, setModelVariant } from '../functions/loaders/VariantFunctions'
