+++ packages/engine/src/xr/XRLightProbeSystem.ts
@@ -34,9 +34,9 @@ import { createEntity } from '@etherealengine/ecs/src/EntityFunctions'
 import { defineSystem } from '@etherealengine/ecs/src/SystemFunctions'
 import { V_000 } from '../common/constants/MathConstants'
 import { EngineRenderer } from '../renderer/WebGLRendererSystem'
+import { DirectionalLightComponent } from '../renderer/components/DirectionalLightComponent'
 import { addObjectToGroup } from '../renderer/components/GroupComponent'
 import { setVisibleComponent } from '../renderer/components/VisibleComponent'
-import { DirectionalLightComponent } from '../scene/components/DirectionalLightComponent'
 import { TransformComponent } from '../transform/components/TransformComponent'
 import { XRState } from './XRState'
 import { XRSystem } from './XRSystem'
